不要なmodelの削除、データベース構造の再設計
