+++ app/models/order.rb
@@ -1,6 +1,3 @@
 class Order < ApplicationRecord
   belongs_to :user
-  belongs_to :schedule
-  belongs_to :stock, optional: true
-
 end
