--- app/models/stock.rb
@@ -1,6 +1,0 @@
-class Stock < ApplicationRecord
-  has_many :stock_details
-  has_many :orders
-
-  belongs_to :schedule
-end
