--- app/models/stock_detail.rb
@@ -1,3 +1,0 @@
-class StockDetail < ApplicationRecord
-  belongs_to :stock
-end
