+++ app/models/ticket.rb
@@ -1,3 +1,4 @@
 class Ticket < ApplicationRecord
+  belongs_to :schedule
   belongs_to :user
 end
