--- db/migrate/20200417041010_create_tickets.rb
@@ -1,11 +1,0 @@
-class CreateTickets < ActiveRecord::Migration[5.2]
-  def change
-    create_table :tickets do |t|
-      t.string :row, null: true
-      t.string :seat_number, null: true
-      t.references :order, null: false, foreign_key: true
-      t.references :user, null: false, foreign_key: true
-      t.timestamps
-    end
-  end
-end
