--- db/migrate/20200418164415_create_stock_details.rb
@@ -1,9 +1,0 @@
-class CreateStockDetails < ActiveRecord::Migration[5.2]
-  def change
-    create_table :stock_details do |t|
-      t.integer :price
-      t.references :stock, null: false, foreign_key: true
-      t.timestamps
-    end
-  end
-end
