--- test/models/stock_detail_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class StockDetailTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end
